Photo modal: floating close (top-right) + fullscreen (photo top-left)
Previously the close `[X]` and fullscreen toggle lived in a Tools
sub-cluster at the right end of the Navigation bar, which pushed
skip-next inward and made the Photo bar look heavier than the
Year / Month bars (where skip-next sits at the right edge).

Both buttons move out of the Navigation bar entirely:
- Close: absolute top-right of the modal Frame, above the
  Navigation via z-index — small enough not to fight skip-next.
- Fullscreen: absolute top-left of the photo area (58px from the
  top of the Frame = below the Navigation bar). Reads as "make
  this photo bigger" rather than "modal chrome".

Both use a shared `<FloatingButton>` style: 34px translucent dark
pill, white icon, hover darkens — visible against any background
(Navigation chrome, matte, dark scrim).

Photo Navigation now matches Year / Month visually: prev/skip on
the left, centre indicator, next/skip on the right, no tools at
the end.
